Ensure that saving nominations is always consistent
Fixes #62

This was a bit of a scary bug, in that it was possible for values in the
ballot to duplicate while saving the formsets.

Rather than do that, we now do this the slow, but very safe way; after
we've confirmed that our form is valid, we wipe out all of the user's
existing nominations, and then bulk save the new ones. This is
duplicative, but it's fast enough for our purposes, and guarantees we
don't need to worry about doubling the data.